refactor: Finally remove all rac code to www app and delete ui package.
