+++ apps/www/app/components/index.ts
@@ -1,3 +1,7 @@
+// React Aria Components
+export * from './(ui)';
+
+// App Components
 export * from './icon/icon';
 export * from './logo/logo';
 export * from './theme-menu/theme-menu';
