+++ apps/www/package.json
@@ -22,15 +22,14 @@
     "@remix-run/node": "^2.7.1",
     "@remix-run/react": "^2.7.1",
     "@remix-run/serve": "^2.7.1",
-    "@tipprunde/ui": "workspace:*",
     "isbot": "^5.1.0",
     "prisma": "^5.10.1",
     "react": "^18.2.0",
+    "react-aria-components": "^1.1.1",
     "react-dom": "^18.2.0",
     "remix-auth": "^3.6.0",
     "remix-auth-totp-dev": "^3.0.0",
     "sonner": "^1.4.0",
-    "tailwind-merge": "^2.2.1",
     "zod": "^3.22.4"
   },
   "devDependencies": {
@@ -44,6 +43,8 @@
     "clsx": "^2.1.0",
     "remix-flat-routes": "^0.6.4",
     "svg-icons-cli": "^0.0.7",
+    "tailwind-merge": "^2.2.1",
+    "tailwind-variants": "^0.2.0",
     "tailwindcss": "^3.4.1",
     "typescript": "^5.3.3",
     "vite": "^5.1.3"
