--- packages/ui/CHANGELOG.md
@@ -1,139 +1,0 @@
-# @tipprunde/ui
-
-## 0.2.4
-
-### Patch Changes
-
-- [`2636326`](https://github.com/haus23/runde.tips/commit/2636326e8e72d5dc3c7d3aaf64ddbd81abaea706) Thanks [@haus23](https://github.com/haus23)! - chore: Update tailwind-variants
-
-- [`2636326`](https://github.com/haus23/runde.tips/commit/2636326e8e72d5dc3c7d3aaf64ddbd81abaea706) Thanks [@haus23](https://github.com/haus23)! - chore: Update svg icons cli
-
-- [`2636326`](https://github.com/haus23/runde.tips/commit/2636326e8e72d5dc3c7d3aaf64ddbd81abaea706) Thanks [@haus23](https://github.com/haus23)! - chore: Upgrade remix. Vite plugin now stable.
-
-- [`2636326`](https://github.com/haus23/runde.tips/commit/2636326e8e72d5dc3c7d3aaf64ddbd81abaea706) Thanks [@haus23](https://github.com/haus23)! - chore: Update typings.
-
-- [`2636326`](https://github.com/haus23/runde.tips/commit/2636326e8e72d5dc3c7d3aaf64ddbd81abaea706) Thanks [@haus23](https://github.com/haus23)! - chore: Update react-aria-components.
-
-## 0.2.3
-
-### Patch Changes
-
-- fix: Add center class.
-
-- refactor: Rethink button colors.
-
-- feat: Add icons for master data nav items.
-
-- fix: Cleaner header and nav styling. Resolves #10
-
-- fix: Hoppla, must keep the icon sources.
-
-## 0.2.2
-
-### Patch Changes
-
-- refactor: Simplify icon usage.
-
-- refactor: Add navlink variants for top- and sidenav
-
-- feat: Add more icons. And ignore source icons now.
-
-- feat: Add icons logout and settings.
-
-- refactor: Use semibold font.
-
-- feat: Add usage of classname prop
-
-- fix: Move something to deps. Icons build script crashes else.
-
-## 0.2.1
-
-### Patch Changes
-
-- [#8](https://github.com/haus23/runde.tips/pull/8) [`2900977`](https://github.com/haus23/runde.tips/commit/29009774b23f55ddce4acbf2f084ad25b2ee7968) Thanks [@haus23](https://github.com/haus23)! - feat: Implement login form logic
-
-- [#8](https://github.com/haus23/runde.tips/pull/8) [`2900977`](https://github.com/haus23/runde.tips/commit/29009774b23f55ddce4acbf2f084ad25b2ee7968) Thanks [@haus23](https://github.com/haus23)! - feat: Add className prop for the input field.
-
-## 0.2.0
-
-### Minor Changes
-
-- feat: Add unstyled textfield component.
-
-- feat: Styled text field.
-
-### Patch Changes
-
-- refactor: Add overlay-arrow.
-
-- feat: Add simple link components.
-
-- refactor: Change focus ring color to accent color.
-
-- fix: Remove border styling for current links. Resolves #4
-
-- feat: First try on the link components and the app header.
-
-- refactor: Align variant naming.
-
-- refactor: Create a new neutral bg color. First usage in popover bg.
-
-## 0.1.3
-
-### Patch Changes
-
-- feat: Finally styled the theme menu.
-
-## 0.1.2
-
-### Patch Changes
-
-- feat: Add unstyled but working menu component.
-
-- fix: Use default cursor on buttons.
-
-- feat: Add check icon.
-
-## 0.1.1
-
-### Patch Changes
-
-- chore: Update deps. Update Remix to v2.6
-
-- feat: Add minimal styled button with variants.
-
-## 0.1.0
-
-### Minor Changes
-
-- feat: Add icon handling support and an icon component.
-
-### Patch Changes
-
-- chore: Add first react aria component.
-
-- feat: Add complete logo implementation.
-
-- chore: Add react aria components and tailwind helpers.
-
-## 0.0.3
-
-### Patch Changes
-
-- fix: Remove unneeded types property.
-
-- chore: Revert back to completely internal package.
-
-## 0.0.2
-
-### Patch Changes
-
-- chore: Add typecheck script.
-
-- chore: Turn ui package into a transpiled package.
-
-## 0.0.1
-
-### Patch Changes
-
-- chore: Create initial internal ui package. No build step by now.
