--- packages/ui/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@tipprunde/ui",
-  "version": "0.2.4",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "typecheck": "tsc"
-  },
-  "dependencies": {
-    "@remix-run/react": "^2.7.1",
-    "react-aria-components": "^1.1.1"
-  },
-  "devDependencies": {
-    "@tipprunde/tailwind-config": "workspace:*",
-    "@tipprunde/typescript-config": "workspace:*",
-    "@types/react": "^18.2.57",
-    "clsx": "^2.1.0",
-    "react": "^18.2.0",
-    "tailwind-merge": "^2.2.1",
-    "tailwind-variants": "^0.2.0",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5.3.3",
-    "vite": "^5.1.3"
-  }
-}
