--- packages/ui/src/index.ts
@@ -1,14 +1,0 @@
-import '@tipprunde/tailwind-config/styles.css';
-
-// React Aria Types
-export type { Key } from 'react-aria-components';
-
-// React Aria Provider / Components
-export { RouterProvider } from 'react-aria-components';
-export { Form } from 'react-aria-components';
-
-// React Aria Components
-export * from './(ui)/button/button';
-export * from './(ui)/menu/menu';
-export * from './(ui)/link/link';
-export * from './(ui)/textfield/textfield';
